backend/requirements.txt
mistralai
python-dotenv
torch
transformers
numpy
scikit-learn
